Player left the game

+++ src/Multiplayer/Server.java
@@ -220,11 +220,13 @@ public class Server implements Runnable{
                         String[] messageSplit = message.split(" ",2);
                         if(clientName != null && lobby.player1 != null && lobby.player2 != null){
                             write.println("StartGame");
+                            lobby.running = true;
                         }
                         for (ConnectionHandler ch : connections){
                             if(ch.clientName != null && ch.lobby.player1 != null && ch.lobby.player2 != null){
                                 if((ch.lobby.player1.equals(messageSplit[1]) || ch.lobby.player2.equals(messageSplit[1])) && !ch.clientName.equals(messageSplit[1])){
                                     ch.write.println("StartGame");
+                                    ch.lobby.running = true;
                                 }
                             }
                         }
@@ -233,6 +235,14 @@ public class Server implements Runnable{
                         for (ConnectionHandler ch : connections){
                             if(!ch.clientName.equals(messageSplit[1]) && (ch.lobby.player1.equals(messageSplit[1]) || ch.lobby.player2.equals(messageSplit[1]))){
                                 ch.write.println("Coordinates" + ' ' + messageSplit[2] + ' ' + messageSplit[3] + ' ' + messageSplit[4] + ' ' + messageSplit[5]);
+                            }
+                        }
+                    }else if(message.startsWith("LeftGame")){
+                        String[] messageSplit = message.split(" ",2);
+                        for (ConnectionHandler ch : connections){
+                            if(!ch.clientName.equals(messageSplit[1]) && (ch.lobby.player1.equals(messageSplit[1]) || ch.lobby.player2.equals(messageSplit[1]))){
+                                mySqlDatabase.deleteLobby(ch.lobby.player1);
+                                ch.write.println("PlayerLeft");
                             }
                         }
                     }
@@ -253,7 +263,11 @@ public class Server implements Runnable{
          */
         public void shutdown() throws SQLException {
             for (ConnectionHandler ch : connections){
-                if(ch.lobby.player2 != null && ch.lobby.player2.equals(clientName) && !ch.clientName.equals(clientName)){
+                if(ch.lobby.running && !ch.clientName.equals(clientName) && (ch.lobby.player1.equals(clientName) || ch.lobby.player2.equals(clientName))){
+                    mySqlDatabase.deleteLobby(ch.lobby.player1);
+                    ch.write.println("PlayerLeft");
+                }
+                if(ch.lobby.player2 != null && ch.lobby.player2.equals(clientName) && !ch.clientName.equals(clientName) && !ch.lobby.running){
                     ch.lobby.player2 = null;
                     this.lobby.player1 = null;
                     this.lobby.player2 = null;
@@ -262,7 +276,7 @@ public class Server implements Runnable{
                     ch.write.println("LeftLobby");
                     break;
                 }
-                if(ch.lobby.player1 != null &&  ch.lobby.player1.equals(clientName)){
+                if(ch.lobby.player1 != null &&  ch.lobby.player1.equals(clientName) && !ch.lobby.running){
                     mySqlDatabase.deleteLobby(clientName);
                     if(!ch.clientName.equals(clientName)){
                         System.out.println(clientName + " left lobby");
